Move distance function from web to util
